Refactor blob storage access and add access checker tool
Removed AccessBlobs.cshtml, BlobDetails.cs, BlobStorageModel.cs,
and refactored BlobStorageController.cs to remove blob access
methods. Updated Index.cshtml to change blob access link and
add access checker link. Added BlobStorageAccessCheckController.cs,
BlobStorageAccessCheckModel.cs, BlogAccessTestResult.cs, and
new Index.cshtml for access checker tool.
