+++ CloudDebugger/Features/BlobStorage/Index.cshtml
@@ -8,7 +8,9 @@
 <br />
 
 
-<a href="/BlobStorage/AccessBlobs">Access blobs</a> (view, create and delete blobs)
+<a href="/BlobStorageEditor/AccessBlobs">Access blobs</a> (view, create and delete blobs)
+<br />
+<a href="/BlobStorageAccessCheck/Index">Access checker</a> (Checks what access do you have to blob storage)
 <br />
 <a href="/BlobStorageDelegationSasToken/Index">Create a user delegation SAS token</a>
 <br />
