+++ CloudDebugger/Features/BlobStorageEditor/AccessBlobs.cshtml
@@ -1,4 +1,4 @@
-﻿@model CloudDebugger.Features.BlobStorage.BlobStorageModel
+﻿@model CloudDebugger.Features.BlobStorageEditor.BlobStorageModel
 @{
     ViewData["Title"] = "Access Blob Storage";
 }
